feat: 1. rename file name 2. fixed SCAN-3944
