+++ src/app/containers/Charts/Difficulty.tsx
