+++ src/app/containers/Charts/HashRate.tsx
