+++ src/app/containers/Charts/TotalSupply.tsx
