+++ src/app/containers/Footer/Language.tsx
@@ -5,6 +5,7 @@ import styled from 'styled-components/macro';
 import Globe from '@zeit-ui/react-icons/globe';
 import { trackEvent } from '../../../utils/ga';
 import { ScanEvent } from '../../../utils/gaConstants';
+import dayjs from 'dayjs';
 
 export function Language() {
   const { i18n } = useTranslation();
@@ -18,6 +19,7 @@ export function Language() {
         label: 'en',
       });
       i18n.changeLanguage('en');
+      dayjs.locale('en');
     } else {
       trackEvent({
         category: ScanEvent.preference.category,
@@ -25,6 +27,7 @@ export function Language() {
         label: 'zh-CN',
       });
       i18n.changeLanguage('zh-CN');
+      dayjs.locale('zh-cn');
     }
   };
   const options = [
