+++ src/app/containers/Header/index.tsx
@@ -23,6 +23,7 @@ import { useGlobalData, GlobalDataType } from 'utils/hooks/useGlobal';
 import { getNetwork, gotoNetwork } from 'utils';
 import { NETWORK_TYPE, NETWORK_TYPES } from 'utils/constants';
 import { Notices } from 'app/containers/Notices/Loadable';
+import dayjs from 'dayjs';
 
 import logo from 'images/logo.svg';
 import logoTest from 'images/logo-test.svg';
@@ -563,7 +564,10 @@ export const Header = memo(() => {
               label: 'en',
             });
             menuClick();
-            return iszh && i18n.changeLanguage('en');
+            if (iszh) {
+              i18n.changeLanguage('en');
+              dayjs.locale('en');
+            }
           },
           isMatchedFn: () => !iszh,
         },
@@ -577,7 +581,10 @@ export const Header = memo(() => {
               label: 'zh-CN',
             });
             menuClick();
-            return !iszh && i18n.changeLanguage('zh-CN');
+            if (!iszh) {
+              i18n.changeLanguage('zh-CN');
+              dayjs.locale('zh-cn');
+            }
           },
           isMatchedFn: () => iszh,
         },
